+++ include/flecs/private/api_types.h
@@ -137,9 +137,9 @@ struct ecs_iter_t {
     ecs_entity_t *entities;       /**< Entity identifiers */
 
     void *param;                  /**< User data (EcsContext or param argument) */
-    float delta_time;             /**< Time elapsed since last frame */
-    float delta_system_time;      /**< Time elapsed since last system invocation */
-    float world_time;             /**< Time elapsed since start of simulation */
+    double delta_time;             /**< Time elapsed since last frame */
+    double delta_system_time;      /**< Time elapsed since last system invocation */
+    double world_time;             /**< Time elapsed since start of simulation */
 
     int32_t frame_offset;         /**< Offset relative to frame */
     int32_t table_offset;         /**< Current active table being processed */
